decimal

+++ MedicalProject/Models/Inventory/AddInventoryCommand.cs
@@ -5,7 +5,7 @@ namespace MedicalProject.Models.Inventory
     public class AddInventoryCommand 
     {
         public string totalPrice { get; set; }
-        public int dong { get; set; }
+        public decimal dong { get; set; }
         public string profit { get; set; }
         public Guid productId { get; set; }
         public PaymentTime? paymentTime { get; set; }
@@ -13,7 +13,7 @@ namespace MedicalProject.Models.Inventory
     public class EditInventoryCommand 
     {
         public string totalPrice { get; set; }
-        public int dong { get; set; }
+        public decimal dong { get; set; }
         public string profit { get; set; }
         public Guid productId { get; set; }
         public PaymentTime? paymentTime { get; set; }
