+++ MedicalProject/Models/SiteEntity/MainPageDto.cs
@@ -5,7 +5,7 @@ namespace MedicalProject.Models.SiteEntity
 {
     public class MainPageDto
     {
-        public int InvestmentNmber { get; set; }
+        public decimal InvestmentNmber { get; set; }
         public List<ProductMainPageQuery> product { get; set; } = new();
     }
     public class ProductMainPageQuery : BaseDto
@@ -15,6 +15,6 @@ namespace MedicalProject.Models.SiteEntity
         public SeoData seoData { get; set; }
         public InventoryDto Inventory { get; set; }
         public ProductStatus Status { get; set; }
-        public int DangRemains { get; set; }
+        public decimal DangRemains { get; set; }
     }
 }
